requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
boto3==1.36.21
botocore==1.36.21
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
-e git+ssh://git@github.com/alekst23/creocortex.git@c4de9b1d6f9a599eb9e0aa232632436b1e07d997#egg=creo
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-aws==0.2.12
langchain-core==0.3.35
langchain-openai==0.3.6
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.27.1
numpy==2.2.3
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.11.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.1
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zstandard==0.23.0
